+++ packages/rax-app/package.json
@@ -2,7 +2,8 @@
   "name": "rax-app",
   "version": "2.1.2",
   "description": "Rax application framework for univesal app.",
-  "main": "src/index.js",
+  "main": "lib/index.js",
+  "module": "es/index.js",
   "author": "Rax Team",
   "repository": {
     "type": "git",
