+++ packages/rax-app/src/router.js
@@ -1,7 +1,30 @@
-// current: { path, visibiltyState }
-const router = {
-  prev: null,
-  current: null
-};
+import { createElement } from 'rax';
+import { withRouter as webSPAWithRouter } from 'rax-use-router';
+import { isMiniApp, isWeChatMiniProgram, isByteDanceMicroApp } from 'universal-env';
 
-export default router;
+let withRouter;
+
+if (isMiniApp || isWeChatMiniProgram || isByteDanceMicroApp) {
+  withRouter = function(Component) {
+    function Wrapper(props) {
+      const history = window.history;
+      return createElement(Component, Object.assign({}, props, {
+        history: history,
+        location: history.location
+      }));
+    };
+    Wrapper.displayName = 'withRouter(' + (Component.displayName || Component.name) + ')';
+    Wrapper.WrappedComponent = Component;
+    return Wrapper;
+  };
+} else {
+  withRouter = webSPAWithRouter;
+}
+
+export default {
+  // prev: { path, visibiltyState }
+  // current: { path, visibiltyState }
+  prev: null,
+  current: null,
+  withRouter
+};
